#8.14 bug fix

+++ src/controllers/videoController.js
@@ -21,21 +21,41 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id }
+    }
+  } = req;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found." })
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); // 403: Forbidden
   }
   return res.render("videos/edit", { pageTitle: `✂ ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
+  const {
+    params: { id },
+    session: {
+      user: { _id }
+    },
+    body: {
+      title, description, hashtags,
+    }
+  } = req;
   const video = await Video.exists({_id: id});
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video Not Found." })
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); // 403: Forbidden
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -85,7 +105,19 @@ export const postUpload = async (req, res) => {
 }
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id }
+    }
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video Not Found." })
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/"); // 403: Forbidden
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 }
